requirements.txt
appdirs==1.4.3
Django==2.0.4
djangorestframework==3.8.2
packaging==17.1
pyparsing==2.2.0
pytz==2018.4
six==1.11.0
psycopg2==2.7.4
django-filter==1.1.0
psycopg2-binary==2.7.4
gunicorn==19.7.1
django-autoslug-iplweb==1.9.4.dev0
pillow==5.1.0
requests==2.18.4
djutils==0.3.2
urllib3==1.21.1
django-elasticsearch-dsl